UI Update
Added placeholder images for the application.  The icon, and app title
are now completely responsive to changing UI.

Added JS that displays the current resolution of the viewing window as
well as how much vertical scrolling has occurred (for debug purposes).

The navigation bar is now sticky.  When the user scrolls past the
header, the navigation bar follows with the user's scroll.

Added mockup login form.  Has no behavior.