static posts, nav update
The static application now contains some static, dummy posts.  When you
hover over a post, the border changes to an orange color.  When you pin
a post, the pin icon in the top-left corner for that post should turn to
orange instead of black.  This feature isn't implemented yet.  As well,
when you receive a notification about  a post you have pinned, the pin
icon on the nav bar will turn to orange.  This feature isn't implemented
yet