Minor UI adjustments, JavaScript
Added a <noscript> tag to be able to detect if javascript is disabled.
if so, we display a warning message.  I put the window statistics to the
bottom left corner of the screen.  Cleaned up the javascript code so
handlers can only be run when the document is ready.  Added a debug
script that lets us toggle between debug modes and how we display debug
output (alerts vs console.log).  When you click on a nav item, its
background changes to dark grey.