Layout changes, ask button, categories drop down
Found some MAJOR flow problems with our document. Fixed the layout and
everything is fine now.  If you click the categories icon on the nav
bar, its menu drops down. very rudimentary- just placeholder code for
starting look and feel.  "Ask a question" button is fixed to bottom
right of content area for now.  We'll make it float with the user later.