Finished Up UI
updated the look and feel.  When you click on a nav item, it hides all
the previously selected nav items.  added upvote/downvote buttons to
comment section.  when you click on a post it brings you to a another
screen

all interactive clicking stuff shows/hides the proper forms and pages.