Login Finished
After logging in successfully, a message is displayed to the user and
he/she is redirected to the home page.  If the user authentication
fails, the password field is cleared and the user remains on the home
page (and of course, an error is displayed).

After logging in, the settings dropdown changes to "logout".  After
logging out, the settings dropdown changes to "login" and the user is
redirected to the login page