Login/Register Validation
When logging in, I use regex to check if the username/password combo is
in the correct form.

As of now, usernames must be alphanumeric (can be mixed case) and can
also contain an underscore.  Usernames must be 6-20 characters long.

Passwords must be 6-30 characters long and can contain alphanumeric
characters and any combination of "!@#$%^&*()".

The app does not send the AJAX request if the regex check fails, and
immediately displays the error message.  Upon successful
login/registration, a message is displayed to the user instead of an
alert()